Move TaskList out of task package
