+++ src/main/java/didier/UI.java
@@ -3,7 +3,6 @@ package didier;
 import didier.exception.DidierException;
 import didier.exception.TaskNumberException;
 import didier.task.Task;
-import didier.task.TaskList;
 
 import java.util.Scanner;
