+++ src/main/java/didier/command/AddCommand.java
@@ -1,6 +1,7 @@
 package didier.command;
 
 import didier.Storage;
+import didier.TaskList;
 import didier.UI;
 import didier.exception.DidierException;
 import didier.task.*;
