+++ src/main/java/didier/command/Command.java
@@ -3,7 +3,7 @@ package didier.command;
 import didier.Storage;
 import didier.UI;
 import didier.exception.DidierException;
-import didier.task.TaskList;
+import didier.TaskList;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, UI ui, Storage storage) throws DidierException;
