+++ src/main/java/didier/command/MarkCommand.java
@@ -4,7 +4,7 @@ import didier.Storage;
 import didier.UI;
 import didier.exception.TaskNumberException;
 import didier.task.Task;
-import didier.task.TaskList;
+import didier.TaskList;
 
 public class MarkCommand extends Command {
