+++ src/test/java/didier/ParserTest.java
@@ -1,2 +1,9 @@
-package didier;public class ParserTest {
+package didier;
+
+import org.junit.jupiter.api.Test;
+
+public class ParserTest {
+
+    @Test
+    public void parse_list_
 }
